5-thread audit: Sys_Error refactor, server info hardening, shift UB, keepalive audit, clc_stringcmd validation
Thread 1 — Sys_Error refactor.  The libretro Sys_Error
historically logged-and-returned (the original Quake
exit() / abort path is impossible in a libretro core),
leaving ~190 call sites silently fall through with
broken state when their failure invariants tripped.
Mirrored Host_Error's setjmp/longjmp pattern:

  - Added jmp_buf sys_abort + sys_abort_set guard +
    sys_error_pending flag in libretro.c.

  - Sys_Error is now void: vsnprintfs into 256-byte
    buffer, logs, sets sys_error_pending, longjmps via
    sys_abort if sys_abort_set.  Falls through to log-
    and-return as last resort for early-init / late-
    shutdown calls without a live setjmp.

  - retro_run wraps Host_Frame in setjmp.  On longjmp,
    skips CL_DecayLights but still presents a dupe
    video frame and runs audio_step so the frontend
    sees steady cadence rather than a stalled core.

  - retro_load_game wraps Host_Init in setjmp.  On
    longjmp, shows error via SET_MESSAGE and returns
    false.

  - sys.h prototype changed to void with no-return
    semantics doc.

  - Fixed 11 `return Sys_Error(...)` callers in host.c
    and wad.c -- converted single-statement
    `if (cond) return Sys_Error(...)` into braced
    `if (cond) { Sys_Error(...); return false; }`.
    The return false is unreachable after the longjmp
    but provides correct fallback in early-init.

Thread 2 — SV_SendServerinfo / precache hardening.

  - PF_precache_model / PF_precache_sound had no
    length cap on QC-supplied names; PR_CheckEmptyString
    only rejected empty / leading-space.  A multi-KB
    QC name would be stored in sv.{model,sound}_
    precache[] and reflected to every connecting
    client via SV_SendServerinfo's MSG_WriteString
    loop, blowing past client->message.maxsize.  Added
    strlen(s) >= MAX_QPATH cap in PR_CheckEmptyString
    (sister to PF_lightstyle's MAX_STYLESTRING bound).

  - SV_SendServerinfo's map title transmission used
    PR_GetString(sv.edicts->v.message) directly with
    no length cap (QC-set during worldspawn).  Buffered
    into a stack-local char[80] via strlcpy first --
    truncates rather than dropping the field.

Thread 3 — MSG_ReadLong << 24 UB.  Three sites had
the same pattern: net_message.data[i] << 24 where
data[i] is byte promoted to int.  For data >= 0x80 the
shifted value >= 0x80000000 exceeds INT_MAX, and
shifting an int into the sign bit is implementation-
defined per C99 6.5.7.  All real compilers wrap to
negative correctly, but the unsigned form is well-
defined.

  - MSG_ReadLong (common.c): unsigned accumulation,
    cast result to int at return.

  - MSG_ReadControlHeader (common.c): same fix.

  - CL_ParseUpdate / CL_ParseClientdata (cl_parse.c):
    `bits |= MSG_ReadByte() << 24` for FITZ extend
    bits -- cast to unsigned before shift.

Thread 4 — CL_KeepaliveMessage timing.  Audited; no
fix needed.  Function early-returns on sv.active
(i.e. local listenserver, ~100% of libretro Quake
usage), so the 5-second host_time cooldown is only
exercised when connected to a remote server -- where
Host_Frame ticks normally and host_time advances as
wall-clock would.  Strengthened existing comment to
record the analysis and note that lastmsg (static)
persisting across maps is fine because host_time is
monotonic since process start.

Thread 5 — Server-side clc_stringcmd validation.
SV_ReadClientMessage's clc_stringcmd whitelist was a
chain of strncasecmp prefix checks.  Functionally
non-exploitable (unknown extended-prefix tokens like
"saywhatever" would match "say" -> ret=1 -> Cmd_
FindCommand fails -> logged at cvar fallthrough), but
the matching logic was wrong in spirit.

  - Replaced with exact-token whitelist matching:
    skip leading whitespace, find token length,
    compare strlen(allowed[i]) == cmd_len AND
    strncasecmp == 0.  Per-handler cmd_source checks
    (verified during Thread D audit) remain as the
    inner perimeter; this whitelist is the outer.

  - Added 1024-byte length cap on s itself to prevent
    Cmd_TokenizeString Z_Malloc storms on hostile-
    large stringcmds (MSG_ReadString bounds at
    COM_STRBUF_LEN, but a 64KB legal command string
    would still blow up Cmd_Tokenize).

Verification:
  gcc -fsyntax-only -DNQ_HACK -Wimplicit-function-
    declaration -Wunused-function clean across 67-file
    build subset.
  Patch applies cleanly to 1fda936 + the still-
    pending extern-deadcvar-cleanup baseline.

+++ common/cl_parse.c
@@ -268,7 +268,21 @@ CL_KeepaliveMessage(void)
  * to prevent the server from timing out the client
  * during long load operations, and "long" measured
  * in frames is the relevant quantity to a libretro
- * core. */
+ * core.
+ *
+ * Note: reaching this point at all requires sv.active
+ * to be false (line 236) -- the libretro core almost
+ * always runs a local listenserver, so the keepalive
+ * is effectively only exercised when connected to a
+ * remote network server, where Host_Frame continues to
+ * tick and host_time advances normally.  In that mode
+ * the 5-second cooldown via host_time behaves exactly
+ * as wall-clock would.
+ *
+ * lastmsg persists across map / server transitions,
+ * but host_time is monotonic since process start
+ * (see round-5 host clock work) so the relative
+ * delta stays sensible. */
     time = host_time;
     if (time - lastmsg < 5)
 	return;
@@ -501,10 +515,16 @@ CL_ParseUpdate(unsigned int bits)
    }
 
    if (cl.protocol == PROTOCOL_VERSION_FITZ) {
+      /* Cast to unsigned before shift: MSG_ReadByte returns
+       * int [0..255], and shifting 0x80..0xff left by 24
+       * pushes into the sign bit (implementation-defined
+       * per C99 6.5.7).  bits is unsigned int so the
+       * subsequent OR is fine; the issue is the shift
+       * itself. */
       if (bits & U_FITZ_EXTEND1)
-         bits |= MSG_ReadByte() << 16;
+         bits |= (unsigned)MSG_ReadByte() << 16;
       if (bits & U_FITZ_EXTEND2)
-         bits |= MSG_ReadByte() << 24;
+         bits |= (unsigned)MSG_ReadByte() << 24;
    }
 
    if (bits & U_LONGENTITY)
@@ -734,10 +754,12 @@ CL_ParseClientdata(void)
     unsigned int bits;
 
     bits = (unsigned short)MSG_ReadShort();
+    /* Cast to unsigned before shift: see CL_ParseUpdate
+     * for the same fix. */
     if (bits & SU_FITZ_EXTEND1)
-	bits |= MSG_ReadByte() << 16;
+	bits |= (unsigned)MSG_ReadByte() << 16;
     if (bits & SU_FITZ_EXTEND2)
-	bits |= MSG_ReadByte() << 24;
+	bits |= (unsigned)MSG_ReadByte() << 24;
 
     if (bits & SU_VIEWHEIGHT)
 	cl.viewheight = MSG_ReadChar();
